# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ZyntraCare API
  description: |
    AI-Powered Healthcare Platform API.
    Provides endpoints for symptom checking, health risk assessment,
    emergency services, hospital management, and AI chat.

    ## Authentication
    Most endpoints require JWT authentication via NextAuth.
    Pass the session token in cookies or Authorization header.

    ## Rate Limiting
    All endpoints are rate-limited:
    - Auth endpoints: 20 requests/minute
    - AI endpoints: 30 requests/minute
    - General API: 60 requests/minute

  version: 1.0.0
  contact:
    name: ZyntraCare Team
    url: https://zyntracare.com
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/subscribe:
    post:
      tags: [Health]
      summary: Manage Subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, plan]
              properties:
                action:
                  type: string
                  enum: [subscribe, upgrade, downgrade, cancel, reactivate]
                plan:
                  type: string
                  enum: [Free, Premium Monthly, Premium Yearly]
      responses:
        '200':
          description: Subscription updated
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie