# src/app/api/subscribe/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { getToken } from 'next-auth/jwt';
import { prisma } from '@/lib/prisma';
import { PLAN_BY_NAME, PLANS } from '@/lib/plans';
import { demoOrders, verifyRazorpaySignature } from '@/lib/razorpay';

async function requireAuth(req: NextRequest) {
  const token = await getToken({ req, secret: process.env.NEXTAUTH_SECRET, raw: false });
  if (!token) {
    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
  }
  return token;
}

export async function GET(req: NextRequest) {
  const token = await requireAuth(req);
  if (token instanceof NextResponse) return token;

  return NextResponse.json({ 
    message: 'Subscription API',
    plans: PLANS.map((plan) => plan.name),
    actions: ['subscribe', 'upgrade', 'downgrade', 'cancel', 'reactivate']
  });
}

export async function POST(req: NextRequest) {
  const token = await requireAuth(req);
  if (token instanceof NextResponse) return token;

  try {
    const body = await req.json();
    const { action, plan } = body;
    const userId = token.id as string;

    switch (action) {
      case 'subscribe':
      case 'upgrade': {
        const planConfig = PLAN_BY_NAME[plan];
        if (!planConfig || planConfig.price === 0) {
          return NextResponse.json({ error: 'Invalid plan' }, { status: 400 });
        }

        const { orderId = '', paymentId = '', signature = '' } = body;
        const isDemoOrder = orderId.startsWith('order_demo_');
        const paymentVerified = isDemoOrder
          ? demoOrders.has(orderId)
          : verifyRazorpaySignature(orderId, paymentId, signature);
        if (!paymentVerified) {
          return NextResponse.json({ error: 'Payment verification failed' }, { status: 403 });
        }
        demoOrders.delete(orderId);

        const startDate = new Date();
        const endDate = new Date();

        if (planConfig.period === 'month') {
          endDate.setMonth(endDate.getMonth() + 1);
        } else {
          endDate.setFullYear(endDate.getFullYear() + 1);
        }
        
        const subscription = await prisma.subscription.upsert({
          where: { userId },
          update: { plan, status: 'active', startDate, endDate },
          create: { userId, plan, status: 'active', startDate, endDate }
        });
        
        return NextResponse.json({
          success: true,
          action: action === 'upgrade' ? 'upgraded' : 'subscribed',
          plan: subscription.plan,
          status: subscription.status,
          startDate: subscription.startDate.toISOString(),
          endDate: subscription.endDate?.toISOString(),
          message: action === 'upgrade' ? `Upgraded to ${plan}` : `Subscribed to ${plan}`
        });
      }
      
      case 'downgrade': {
        const currentSub = await prisma.subscription.findUnique({ where: { userId } });
        if (!currentSub) {
          return NextResponse.json({ error: 'No active subscription' }, { status: 400 });
        }
        
        const endDate = new Date();
        endDate.setFullYear(endDate.getFullYear() + 100);
        
        const subscription = await prisma.subscription.update({
          where: { userId },
          data: { plan: 'Free', status: 'active', endDate }
        });
        
        return NextResponse.json({
          success: true,
          action: 'downgraded',
          plan: 'Free',
          status: 'active',
          message: 'Downgraded to Free plan'
        });
      }
      
      case 'cancel': {
        const currentSub = await prisma.subscription.findUnique({ where: { userId } });
        if (!currentSub) {
          return NextResponse.json({ error: 'No active subscription' }, { status: 400 });
        }
        
        const subscription = await prisma.subscription.update({
          where: { userId },
          data: { status: 'cancelled' }
        });
        
        return NextResponse.json({
          success: true,
          action: 'cancelled',
          status: 'cancelled',
          currentPlan: currentSub.plan,
          message: 'Subscription cancelled'
        });
      }
      
      case 'reactivate': {
        const currentSub = await prisma.subscription.findUnique({ where: { userId } });
        if (!currentSub || currentSub.status !== 'cancelled') {
          return NextResponse.json({ error: 'No cancelled subscription to reactivate' }, { status: 400 });
        }
        
        const subscription = await prisma.subscription.update({
          where: { userId },
          data: { status: 'active' }
        });
        
        return NextResponse.json({
          success: true,
          action: 'reactivated',
          status: 'active',
          message: 'Subscription reactivated'
        });
      }
      
      default:
        return NextResponse.json({ error: 'Invalid action' }, { status: 400 });
    }
  } catch (error) {
    console.error('Subscription error:', error);
    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
  }
}